bluez: fix bluetooth.service not enabled by default [1/1]
PD#SWPL-149207

Problem:
bluetooth.service is disabled due to the global `systemd_auto_enabled`
is set to `disabled`.
In the yocto sdk, bluetooth.service is triggered as the dependency of
wpeframework. But in yocto-bsp it is not started as expected.

Solution:
Fix to disable bluez.service only which will be triggered by the udev
script.

Verify:
t5w

Change-Id: I88b1c66fa775c4329ac521d79522384f998472fa
Signed-off-by: Jun Zhang <jun.zhang@amlogic.com>